Refactor tls-certificates integration test to its own file
